:sparkles: Separated TinkerGraph from abstracted Gremlin impl
