+++ build.gradle
@@ -16,9 +16,10 @@ repositories {
 dependencies {
     implementation 'org.apache.logging.log4j:log4j-core:2.8.2'
     implementation 'org.apache.logging.log4j:log4j-slf4j-impl:2.8.2'
-    implementation 'org.apache.tinkerpop:gremlin-core:3.4.5'
-    implementation 'org.apache.tinkerpop:tinkergraph-gremlin:3.4.5'
+    implementation 'org.apache.tinkerpop:gremlin-core:3.4.6'
+    implementation 'org.apache.tinkerpop:tinkergraph-gremlin:3.4.6'
     implementation 'org.janusgraph:janusgraph-core:0.5.1'
+    implementation 'org.janusgraph:janusgraph-inmemory:0.5.1'
     implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8'
     testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.4.0'
 }
